+++ src/components/Pages/Dashboard/Dashboard.tsx
@@ -1,15 +1,15 @@
-import { MaxCard } from "../../Global/Card/MaxCard";
-import { UsersInfo } from "../../Layouts/UsersInfo";
+import { MaxCard } from "../../Global/Card/MaxCard/MaxCard";
+import { UsersInfo } from "./UsersInfo/UsersInfo";
 import { ReviewUsers } from "./ReviewUsers/ReviewUsers";
 
 export const Dashboard = () => {
 	return (
-		<MaxCard className="w-full ">
-			<div className="flex ">
-				<div className="">
+		<MaxCard className="w-full bg-red-400 ">
+			<div className="flex">
+				<div className="w-full bg-yellow-300">
 					<UsersInfo />
 				</div>
-				<div>
+				<div className="w-full">
 					<ReviewUsers />
 				</div>
 			</div>
