+++ src/components/Pages/Dashboard/ReviewUsers/Active/Active.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Active = () => {
+	return <div>Active</div>;
+};
