+++ src/components/Pages/Dashboard/ReviewUsers/AllUsers/AllUsers.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const AllUsers = ( ) => {
+	return <div>Users</div>;
+};
