+++ src/components/Pages/Dashboard/ReviewUsers/ReviewUsers.tsx
@@ -1,5 +1,13 @@
-import { Card } from "../../../Global/Card/DefaultCard/Card";
+import { Card } from "../../../Global/Card/Card";
+import { Active } from "./Active/Active";
+import { AllUsers } from "./AllUsers/AllUsers";
 
 export const ReviewUsers = () => {
-	return <Card>ReviewUsers</Card>;
+	return (
+		<Card className="w-1/2">
+			{" "}
+			<Active />
+			<AllUsers />
+		</Card>
+	);
 };
