+++ src/components/Pages/Dashboard/UsersInfo/ActiveUsers/ActiveUsers.tsx
@@ -1,0 +1,8 @@
+import { UserCard } from "../../../../Global/Card/UsersCard/UsersCard";
+import { PiUserGearLight } from "react-icons/pi";
+
+export const ActiveUsers = () => {
+	return (
+		<UserCard info="Active Users" icon={<PiUserGearLight />} percent={`17%`} />
+	);
+};
