+++ src/components/Pages/Dashboard/UsersInfo/UsersInfo.tsx
@@ -1,0 +1,13 @@
+import { ActiveUsers } from "./ActiveUsers/ActiveUsers";
+import { AllUsers } from "./AllUsers/AllUsers";
+import { UsersTotalAmountOnBet } from "./UsersTotalAmountOnBet/UsersTotalAmountOnBet";
+
+export const UsersInfo = () => {
+	return (
+		<div className="flex  sm:flex-col lg:flex-row sm:items-center sm:justify-center lg:justify-start lg:items-start space-x-3 w-full bg-teal-400">
+			<AllUsers />
+			<ActiveUsers />
+			<UsersTotalAmountOnBet />
+		</div>
+	);
+};
