+++ src/components/Pages/Dashboard/UsersInfo/UsersTotalAmountOnBet/UsersTotalAmountOnBet.tsx
@@ -1,0 +1,12 @@
+import { AiTwotoneWallet } from "react-icons/ai";
+import { UserCard } from "../../../../Global/Card/UsersCard/UsersCard";
+
+export const UsersTotalAmountOnBet = () => {
+	return (
+		<UserCard
+			info="Total User Bet"
+			icon={<AiTwotoneWallet />}
+			percent={`17%`}
+		/>
+	);
+};
